move ApiEvaluate to api_evaluation.py
