+++ app/README.md
@@ -24,6 +24,5 @@ $ cd app
 $ sh drucker-grpc-proto/run_codegen.sh
 $ cp drucker-grpc-proto/protobuf/drucker_pb2.py .
 $ cp drucker-grpc-proto/protobuf/drucker_pb2_grpc.py .
-$ python -m unittest test/test_api_service.py
 $ python -m unittest test/test_api_evaluation.py
 ```
