+++ app/apis/api_service.py
@@ -1,11 +1,8 @@
 import datetime
-import json
 
 from flask_restplus import Namespace, fields, Resource, reqparse
 
-from app import logger
-from models import db, Kubernetes, Application, Service, EvaluationResult, Evaluation
-from core.drucker_dashboard_client import DruckerDashboardClient
+from models import db, Kubernetes, Application, Service
 
 from apis.common import DatetimeToTimestamp
 from apis.api_kubernetes import update_dbs_kubernetes, switch_drucker_service_model_assignment
@@ -185,57 +182,4 @@ class ApiApplicationIdServiceId(Resource):
         response_body = {"status": True, "message": "Success."}
         db.session.commit()
         db.session.close()
-        return response_body
-
-@srv_info_namespace.route('/<int:application_id>/services/<int:service_id>/evaluate')
-class ApiEvaluate(Resource):
-    eval_parser = reqparse.RequestParser()
-    eval_parser.add_argument('evaluation_id', location='form', type=int, required=False)
-    eval_parser.add_argument('overwrite', location='form', type=bool, required=False)
-
-    @srv_info_namespace.expect(eval_parser)
-    def post(self, application_id:int, service_id:int):
-        """evaluate"""
-        args = self.eval_parser.parse_args()
-        eval_id = args.get('evaluation_id', None)
-        if eval_id:
-            eobj = Evaluation.query.filter_by(
-                application_id=application_id,
-                evaluation_id=eval_id).first_or_404()
-        else:
-            # if evaluation_id is not given, use the lastest one.
-            eobj = Evaluation.query\
-                .filter_by(application_id=application_id)\
-                .order_by(Evaluation.register_date.desc()).first_or_404()
-
-        sobj = Service.query.filter_by(
-            application_id=application_id,
-            service_id=service_id).first_or_404()
-
-        robj = db.session.query(EvaluationResult)\
-            .filter(EvaluationResult.model_id == sobj.model_id,
-                    EvaluationResult.evaluation_id == eobj.evaluation_id).one_or_none()
-        if robj is not None and args.get('overwrite', False):
-            return robj.result
-
-        eval_result_path = "eval-result-{0:%Y%m%d%H%M%S}.txt".format(datetime.datetime.utcnow())
-        drucker_dashboard_application = DruckerDashboardClient(logger=logger, host=sobj.host)
-        response_body = drucker_dashboard_application.run_evaluate_model(eobj.data_path, eval_result_path)
-
-        if response_body['status']:
-            result = json.dumps(response_body)
-            if robj is None:
-                robj = EvaluationResult(model_id=sobj.model_id,
-                                        data_path=eval_result_path,
-                                        evaluation_id=eobj.evaluation_id,
-                                        result=result)
-                db.session.add(robj)
-            else:
-                robj.data_path = eval_result_path
-                robj.result = result
-            db.session.flush()
-            response_body = robj.result
-            db.session.commit()
-            db.session.close()
-
         return response_body
